BitmapCalculator: fix use of a recycled bitmap
This commit fixes this warning message which happens with the HSV
color picker:

 "Called getWidth() on a recycle()'d bitmap! This is undefined behavior!"

It is likely due to the fact that createScaledBitmap() will usually
return a new bitmap, but will return the source bitmap if the dimensions
are the same.

This commit makes sure that scaleBitmap() will always return a new
bitmap.