quickfix for occasional color reset to white on recomposition
On rare occasions, (was seen using a color picker in an animated
widget) the color picker will reset the selected color to white after a
recomposition.

This seems to be due to onSizeChanged() being called, even though the
color picker did not really change size.

For the HSV color picker it is possible to re-select automatically the
previously picked color, since the coordinates can be calculated. For
the ImageColorPicker, a more elaborate fix would be needed.